Extract set mines and adjacent numbers to two query functions
