+++ jest.config.ts
@@ -3,6 +3,7 @@ import { createDefaultPreset, JestConfigWithTsJest } from "ts-jest";
 const jestConfig: JestConfigWithTsJest = {
   verbose: true,
   rootDir: "src",
+  resolver: "ts-jest-resolver",
   ...createDefaultPreset(),
 };
