+++ package.json
@@ -24,6 +24,7 @@
     "jest": "^29.7.0",
     "serve": "^14.2.4",
     "ts-jest": "^29.2.6",
+    "ts-jest-resolver": "^2.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.3"
   }
