+++ src/board/generateBoard.ts
@@ -1,6 +1,5 @@
 import { Board } from "../data/type.js";
 import { Cell } from "../data/type.js";
-import { getAdjacentMinesTotal } from "./getAdjacentMines.js";
 
 export const generateBoard = (boardDimension: number): Board => {
   if (boardDimension < 5) {
