+++ src/index.ts
@@ -1,9 +1,18 @@
 import { generateBoard } from "./board/generateBoard.js";
-import { setMines } from "./board/setMines.js";
+import { generateBoardWithMines } from "./board/generateBoardWithMines.js";
+import { generateCompleteBoard } from "./board/generateCompleteBoard.js";
+import {
+  getAdjacentCellsPositions,
+  getAdjacentMinesTotal,
+} from "./board/getAdjacentMines.js";
 import { getRandomPositions } from "./board/getRandomPositions.js";
 import { renderBoard } from "./ui/renderBoard.js";
 
 export const board = generateBoard(8);
+
 getRandomPositions(board);
-setMines();
-renderBoard(board);
+export const minesBoard = generateBoardWithMines(board);
+
+export const completeBoard = generateCompleteBoard(minesBoard);
+
+renderBoard(completeBoard);
